requirements.txt
streamlit
newspaper4k